add aircraft type

+++ PerfomanceCalculator/src/com/cdbt/service/OBSCalParmsService.java
@@ -11,16 +11,17 @@ public class OBSCalParmsService {
 		} else {
 			tree = 0.0D;
 		}
-		double height = CommonUtils.strToDouble(oBSAlt).doubleValue() - CommonUtils.strToDouble(rwyEndAlt).doubleValue();
+		double height = CommonUtils.strToDouble(oBSAlt) - CommonUtils.strToDouble(rwyEndAlt);
 		double corrHeight = height + tree
-				+ CommonUtils.strToDouble(turnLength).doubleValue() * CommonUtils.strToDouble(gradLoss).doubleValue() / 100.0D;
-		double totalLength = CommonUtils.strToDouble(lineLength).doubleValue() + CommonUtils.strToDouble(turnLength).doubleValue();
+				+ CommonUtils.strToDouble(turnLength) * CommonUtils.strToDouble(gradLoss) / 100.0D;
+		double totalLength = CommonUtils.strToDouble(lineLength) + CommonUtils.strToDouble(turnLength);
 		double obsGrad = corrHeight * 100.0D / totalLength;
 		double obsGradWithoutTree = (corrHeight - tree) * 100.0D / totalLength;
 		double obsGradResult = obsGrad + 0.8D;
 		double obsGradWithoutTreeResult = obsGradWithoutTree + 0.8D;
 
-		String[] result = { String.valueOf(num), oBSAlt, CommonUtils.doubleToStr(Double.valueOf(height), Integer.valueOf(1)),
+		String[] result = { String.valueOf(num), oBSAlt,
+				CommonUtils.doubleToStr(Double.valueOf(height), Integer.valueOf(1)),
 				CommonUtils.doubleToStr(Double.valueOf(corrHeight), Integer.valueOf(1)),
 				CommonUtils.doubleToStr(Double.valueOf(totalLength), Integer.valueOf(0)),
 				CommonUtils.doubleToStr(Double.valueOf(obsGrad), Integer.valueOf(2)),
